requirements.txt
aws-cdk-lib==2.151.0
constructs>=10.0.0,<11.0.0
boto3>=1.26
aws_solutions_constructs.aws_lambda_dynamodb>=2.39.0